Started fixes
Stetting up for base functionalities
