--- Assets/Scenes/ChampionSelect.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c2eba31177f366c4ba6a45b118e8f402
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
