+++ Assets/Scenes/Menu.unity
@@ -2540,7 +2540,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -323, y: -12}
+  m_AnchoredPosition: {x: -329.99997, y: -107.05}
   m_SizeDelta: {x: 380.2, y: 359.4}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1450690853
