--- Assets/Scripts/Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7622ed6e0c259f64bacf6e92ef5f4552
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
