+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,15 +6,18 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 0
-    path: Assets/Scenes/SampleScene.unity
-    guid: 99c9720ab356a0642a771bea13969a05
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 0
     path: 
     guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/Menu.unity
     guid: 0f933233baa11724eb56a1143340a85f
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/ChampionSelect.unity
     guid: c2eba31177f366c4ba6a45b118e8f402
+  - enabled: 1
+    path: Assets/Scenes/Lobby.unity
+    guid: 8f60522ea258e7446bdf99161453e092
   m_configObjects: {}
